Fix some Ember Data issues
In the infinite scroll implementation we were depending on some undefined behavior that no longer works in newer Ember Data versions.

Calling `setObjects` to update the list of data and force a rerender no longer works. We now store the list in a tracked property which we update when needed to force a rerender.