Clean up the VoteOverview component
This removes the d3 based code which is no longer used since we're now using a plain svg chart.

This has the extra benefit that we're not importing d3 anywhere else but the map, so Embroider can move it from the main bundle to the map route bundle reducing the initial payload a bit.